requirements.txt
fastapi>=0.133.0
psycopg2-binary>=2.9.11
pydantic>=2.12.5
sqlalchemy>=2.0.46
uvicorn>=0.41.0
python-dotenv>=1.0.1

